edit add & delete & edit userLessons Controller and view and added 403 error page and gate all controller and view

+++ resources/views/backend/admin/userLessons/index.blade.php
@@ -21,28 +21,27 @@
             @endif
 
         </section>
+        <h4>{{ $title }}</h4>
+        <h5>{{ $userTitle }}</h5>
         <table class="table table-hover table-bordered">
-            <thead class="bg-success" style="font-size: 15px ; font-family: Tahoma; background-color: #67b168;text-align: center">
+            <thead class="bg-success"
+                   style="font-size: 15px ; font-family: Tahoma; background-color: #67b168;text-align: center">
             <td>
                 <label style="color: white"> ردیف </label>
             </td>
             <td>
-                <label style="color: white"> عنوان پایه </label>
+                <label style="color: white">دروس</label>
             </td>
 
-
-            <td colspan="3" style="text-align: center">
-                <label style="color: white">ویرایش</label>
-            </td>
-
+            @if(\Illuminate\Support\Facades\Gate::check('Admin') || \Illuminate\Support\Facades\Gate::check('Teacher'))
+                <td colspan="3" style="text-align: center">
+                    <label style="color: white">ویرایش</label>
+                </td>
+            @endif
             </thead>
             <tbody>
 
             @foreach($userLessons as $item)
-
-
-
-
                 <tr style="text-align: justify">
                     <td>
                         <label style="color: black">{{ ++$row }}</label>
@@ -50,17 +49,19 @@
                     <td>
                         <label style="color: black">{{ $item->title }} </label>
                     </td>
+                    @if(\Illuminate\Support\Facades\Gate::check('Admin') || \Illuminate\Support\Facades\Gate::check('Teacher'))
+                        <td style="text-align: center">
 
-                    <td style="text-align: center">
-
-                        <a href="{{route('userLessons.edit',['lesson_id' => $item->id, 'teacher_id' => $teacher_id])}}"><input type="button" class="btn btn-info" style="font-size: 15px;font-family: Tahoma" value="ویرایش"></a>
-                    </td>
-
-                    <td style="text-align: center">
-                        {!! Form::open(['route' => ['userLessons.destroy', ['lesson_id' => $item->id, 'teacher_id' => $teacher_id] ],'method' => 'delete']) !!}
-                        {!! Form::submit('حذف', ['class' => 'btn btn-danger']) !!}
-                        {!! Form::close() !!}
-                    </td>
+                            <a href="{{route('userLessons.edit',['lesson_id' => $item->id, 'teacher_id' => $teacher_id])}}"><input
+                                    type="button" class="btn btn-info" style="font-size: 15px;font-family: Tahoma"
+                                    value="ویرایش"></a>
+                        </td>
+                        <td style="text-align: center">
+                            {!! Form::open(['route' => ['userLessons.destroy', ['lesson_id' => $item->id, 'teacher_id' => $teacher_id] ],'method' => 'delete']) !!}
+                            {!! Form::submit('حذف', ['class' => 'btn btn-danger']) !!}
+                            {!! Form::close() !!}
+                        </td>
+                    @endif
 
                 </tr>
             @endforeach
@@ -68,11 +69,14 @@
 
 
         </table>
-{{--        <span style="float:right">{{$userLessons->links()}}</span>--}}
-        <section class="form-group">
-            <td style="text-align: center"><a href="{{route('userLessons.create',$teacher_id)}}"><input type="button" class="form-control btn btn-info"  style="font-size: 15px;font-family: Tahoma"value="صفحه درج "></a></td>
-        </section>
-
+        {{--        <span style="float:right">{{$userLessons->links()}}</span>--}}
+        @if(\Illuminate\Support\Facades\Gate::check('Admin') || \Illuminate\Support\Facades\Gate::check('Teacher'))
+            <section class="form-group">
+                <td style="text-align: center"><a href="{{route('userLessons.create',$teacher_id)}}"><input
+                            type="button" class="form-control btn btn-info" style="font-size: 15px;font-family: Tahoma"
+                            value="صفحه درج "></a></td>
+            </section>
+        @endif
     </section>
 @endsection
